requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
async-timeout==5.0.1
attrs==25.3.0
certifi==2025.1.31
charset-normalizer==3.4.1
dateparser==1.2.1
filelock==3.13.1
frozenlist==1.5.0
fsspec==2024.6.1
greenlet==3.1.1
idna==3.10
Jinja2==3.1.4
MarkupSafe==2.1.5
mpmath==1.3.0
multidict==6.2.0
networkx==3.3
numpy==2.2.4
pandas==2.2.3
pillow==11.0.0
propcache==0.3.1
pycryptodome==3.22.0
python-binance==1.0.28
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
regex==2024.11.6
requests==2.32.3
six==1.17.0
SQLAlchemy==2.0.40
sympy==1.13.1
triton==3.2.0
typing_extensions==4.13.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.3.0
websockets==15.0.1
yarl==1.18.3
